Increase buf_rec in RPUPacket to 2 bytes
To accomodate the larger number of records which
can be collected.

+++ src/RPUcomm.cpp
@@ -126,7 +126,7 @@ bool RPUComm::RX_Error(char * error, uint8_t buffer_size)
 void RPUPacket::setBoardId(uint16_t id)        { board_id_    = id; }
 void RPUPacket::setState(uint8_t state)        { state_       = state; }
 void RPUPacket::setWdtCount(uint8_t count)     { wdt_count_   = count; }
-void RPUPacket::setBufferedRecords(uint8_t count) { buffered_records_ = count; }
+void RPUPacket::setBufferedRecords(uint16_t count) { buffered_records_ = count; }
 void RPUPacket::setVin(float volts)            { vin_raw_     = (uint8_t)constrain((int)(volts * 10.0f), 0, 255); }
 void RPUPacket::setV5V(float volts)            { v5v_raw_     = (uint8_t)constrain((int)(volts * 10.0f), 0, 255); }
 void RPUPacket::setBatV(float volts)           { bat_v_raw_   = (uint8_t)constrain((int)(volts * 10.0f), 0, 255); }
@@ -165,7 +165,7 @@ bool RPUPacket::encode(uint8_t * buf, size_t buf_size) const
     bsw.write_unchecked<uint16_t>(board_id_,        RPU_PKT_ID_BITS);
     bsw.write_unchecked<uint8_t> (state_,           RPU_PKT_STATE_BITS);
     bsw.write_unchecked<uint8_t> (wdt_count_,       RPU_PKT_WDT_BITS);
-    bsw.write_unchecked<uint8_t> (buffered_records_, RPU_PKT_BUFREC_BITS);
+    bsw.write_unchecked<uint16_t>(buffered_records_, RPU_PKT_BUFREC_BITS);
     bsw.write_unchecked<uint8_t> (vin_raw_,         RPU_PKT_VIN_BITS);
     bsw.write_unchecked<uint8_t> (v5v_raw_,         RPU_PKT_V5_BITS);
     bsw.write_unchecked<uint8_t> (bat_v_raw_,       RPU_PKT_BATV_BITS);
@@ -203,7 +203,7 @@ bool RPUPacket::decode(const uint8_t * buf, size_t buf_size)
     board_id_     = bsr.read_unchecked<uint16_t>(RPU_PKT_ID_BITS);
     state_        = bsr.read_unchecked<uint8_t> (RPU_PKT_STATE_BITS);
     wdt_count_    = bsr.read_unchecked<uint8_t> (RPU_PKT_WDT_BITS);
-    buffered_records_ = bsr.read_unchecked<uint8_t> (RPU_PKT_BUFREC_BITS);
+    buffered_records_ = bsr.read_unchecked<uint16_t>(RPU_PKT_BUFREC_BITS);
     vin_raw_      = bsr.read_unchecked<uint8_t> (RPU_PKT_VIN_BITS);
     v5v_raw_      = bsr.read_unchecked<uint8_t> (RPU_PKT_V5_BITS);
     bat_v_raw_    = bsr.read_unchecked<uint8_t> (RPU_PKT_BATV_BITS);
